Feat: Directory 구조 변경
