--- src/main/java/gift/annotation/AuthUser.java
@@ -1,11 +1,0 @@
-package gift.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PARAMETER})
-public @interface AuthUser {
-}
