--- src/main/java/gift/domain/UserRole.java
@@ -1,5 +1,0 @@
-package gift.domain;
-
-public enum UserRole {
-    NORMAL, ADMIN
-}
