--- src/main/java/gift/dto/request/LoginRequest.java
@@ -1,10 +1,0 @@
-package gift.dto.request;
-
-import jakarta.validation.constraints.Email;
-
-public record LoginRequest(
-        @Email
-        String email,
-        String password
-) {
-}
