--- src/main/java/gift/dto/response/TokenResponse.java
@@ -1,9 +1,0 @@
-package gift.dto.response;
-
-public record TokenResponse(
-        String Token
-) {
-    public static TokenResponse from(String token) {
-        return new TokenResponse(token);
-    }
-}
