--- src/main/java/gift/dto/response/UserResponse.java
@@ -1,15 +1,0 @@
-package gift.dto.response;
-
-import gift.entity.User;
-
-public record UserResponse(
-        Long id,
-        String email
-) {
-    public static UserResponse from(User user) {
-        return new UserResponse(
-            user.getId(),
-            user.getEmail()
-        );
-    }
-}
