--- src/main/java/gift/exception/GlobalExceptionHandler.java
@@ -1,33 +1,0 @@
-package gift.exception;
-
-import gift.exception.token.TokenTypeException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.security.SignatureException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(value = MethodArgumentNotValidException.class)
-    public ResponseEntity<?> handleException(MethodArgumentNotValidException e) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
-                e.getBindingResult()
-                        .getAllErrors()
-                        .getFirst()
-                        .getDefaultMessage()
-        );
-    }
-
-    @ExceptionHandler(value = SignatureException.class)
-    public ResponseEntity<?> handleException(SignatureException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
-    }
-
-    @ExceptionHandler(value = TokenTypeException.class)
-    public ResponseEntity<?> handleException(TokenTypeException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
-    }
-}
