--- src/main/java/gift/exception/InValidSpecialCharException.java
@@ -1,7 +1,0 @@
-package gift.exception;
-
-public class InValidSpecialCharException extends RuntimeException {
-    public InValidSpecialCharException(String message) {
-        super(message);
-    }
-}
