--- src/main/java/gift/exception/gift/NeedAcceptException.java
@@ -1,7 +1,0 @@
-package gift.exception.gift;
-
-public class NeedAcceptException extends RuntimeException {
-    public NeedAcceptException(String message) {
-        super(message);
-    }
-}
