--- src/main/java/gift/exception/gift/NoGiftException.java
@@ -1,7 +1,0 @@
-package gift.exception.gift;
-
-public class NoGiftException extends RuntimeException {
-    public NoGiftException(String message) {
-        super(message);
-    }
-}
