--- src/main/java/gift/exception/gift/NoValueException.java
@@ -1,7 +1,0 @@
-package gift.exception.gift;
-
-public class NoValueException extends RuntimeException {
-    public NoValueException(String message) {
-        super(message);
-    }
-}
