--- src/main/java/gift/exception/token/NoTokenException.java
@@ -1,7 +1,0 @@
-package gift.exception.token;
-
-public class NoTokenException extends RuntimeException {
-    public NoTokenException(String message) {
-        super(message);
-    }
-}
