--- src/main/java/gift/exception/token/TokenExpiredException.java
@@ -1,7 +1,0 @@
-package gift.exception.token;
-
-public class TokenExpiredException extends RuntimeException {
-    public TokenExpiredException(String message) {
-        super(message);
-    }
-}
