--- src/main/java/gift/exception/token/TokenTypeException.java
@@ -1,7 +1,0 @@
-package gift.exception.token;
-
-public class TokenTypeException extends RuntimeException {
-    public TokenTypeException(String message) {
-        super(message);
-    }
-}
