--- src/main/java/gift/exception/user/InValidPasswordException.java
@@ -1,7 +1,0 @@
-package gift.exception.user;
-
-public class InValidPasswordException extends RuntimeException {
-    public InValidPasswordException(String message) {
-        super(message);
-    }
-}
