--- src/main/java/gift/exception/user/NoUserException.java
@@ -1,7 +1,0 @@
-package gift.exception.user;
-
-public class NoUserException extends RuntimeException {
-    public NoUserException(String message) {
-        super(message);
-    }
-}
