+++ src/main/java/gift/shared/annotation/AuthUser.java
@@ -1,0 +1,11 @@
+package gift.shared.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.PARAMETER})
+public @interface AuthUser {
+}
