+++ src/main/java/gift/shared/auth/dto/request/LoginRequest.java
@@ -1,0 +1,10 @@
+package gift.shared.auth.dto.request;
+
+import jakarta.validation.constraints.Email;
+
+public record LoginRequest(
+        @Email
+        String email,
+        String password
+) {
+}
