+++ src/main/java/gift/shared/auth/dto/request/SignUpRequest.java
@@ -1,0 +1,14 @@
+package gift.shared.auth.dto.request;
+
+import gift.user.entity.User;
+import jakarta.validation.constraints.Email;
+
+public record SignUpRequest(
+        @Email
+        String email,
+        String password
+) {
+        public User toEntity() {
+                return new User(this.email(), this.password());
+        }
+}
