+++ src/main/java/gift/shared/auth/dto/response/TokenResponse.java
@@ -1,0 +1,9 @@
+package gift.shared.auth.dto.response;
+
+public record TokenResponse(
+        String Token
+) {
+    public static TokenResponse from(String token) {
+        return new TokenResponse(token);
+    }
+}
