+++ src/main/java/gift/shared/domain/UserRole.java
@@ -1,0 +1,5 @@
+package gift.shared.domain;
+
+public enum UserRole {
+    NORMAL, ADMIN
+}
