+++ src/main/java/gift/shared/exception/product/InValidSpecialCharException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.product;
+
+public class InValidSpecialCharException extends RuntimeException {
+    public InValidSpecialCharException(String message) {
+        super(message);
+    }
+}
