+++ src/main/java/gift/shared/exception/product/NeedAcceptException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.product;
+
+public class NeedAcceptException extends RuntimeException {
+    public NeedAcceptException(String message) {
+        super(message);
+    }
+}
