+++ src/main/java/gift/shared/exception/product/NoProductException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.product;
+
+public class NoProductException extends RuntimeException {
+    public NoProductException(String message) {
+        super(message);
+    }
+}
