+++ src/main/java/gift/shared/exception/product/NoValueException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.product;
+
+public class NoValueException extends RuntimeException {
+    public NoValueException(String message) {
+        super(message);
+    }
+}
