+++ src/main/java/gift/shared/exception/token/NoTokenException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.token;
+
+public class NoTokenException extends RuntimeException {
+    public NoTokenException(String message) {
+        super(message);
+    }
+}
