+++ src/main/java/gift/shared/exception/token/TokenExpiredException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.token;
+
+public class TokenExpiredException extends RuntimeException {
+    public TokenExpiredException(String message) {
+        super(message);
+    }
+}
