+++ src/main/java/gift/shared/exception/token/TokenTypeException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.token;
+
+public class TokenTypeException extends RuntimeException {
+    public TokenTypeException(String message) {
+        super(message);
+    }
+}
