+++ src/main/java/gift/shared/exception/user/InValidPasswordException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.user;
+
+public class InValidPasswordException extends RuntimeException {
+    public InValidPasswordException(String message) {
+        super(message);
+    }
+}
