+++ src/main/java/gift/shared/exception/user/NoUserException.java
@@ -1,0 +1,7 @@
+package gift.shared.exception.user;
+
+public class NoUserException extends RuntimeException {
+    public NoUserException(String message) {
+        super(message);
+    }
+}
