+++ src/main/java/gift/shared/token/service/TokenService.java
@@ -1,0 +1,65 @@
+package gift.shared.token.service;
+
+import gift.shared.exception.token.TokenExpiredException;
+import gift.user.entity.User;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+import java.util.function.Function;
+
+@Service
+public class TokenService {
+    @Value("${jwt-secret-key}")
+    private String secretKey;
+
+    @Value("${jwt-access-token-expire-time}")
+    private Integer accessTokenTTL;
+
+    public boolean isTokenExpired(String token) throws TokenExpiredException {
+        return extractExpiration(token).before(new Date());
+    }
+
+    public String generateToken(User user) {
+        return Jwts.builder()
+                .claim("id", user.getId())
+                .claim("email", user.getEmail())
+                .claim("role", user.getRole())
+                .issuedAt(new Date(System.currentTimeMillis()))
+                .expiration(new Date(System.currentTimeMillis() + accessTokenTTL))
+                .signWith(getSigningKey())
+                .compact();
+    }
+
+    private SecretKey getSigningKey(){
+        byte[] bytes = Decoders.BASE64.decode(secretKey);
+        return Keys.hmacShaKeyFor(bytes);
+    }
+
+    private Date extractExpiration(String token) {
+        return extractClaim(token, Claims::getExpiration);
+    }
+
+    public Long extractId(String token){
+        return extractClaim(token, claims -> claims.get("id", Long.class));
+    }
+
+    private <T> T extractClaim(String token, Function<Claims, T> claimsResolver){
+        final Claims claims = extractAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    private Claims extractAllClaims(String token){
+        return Jwts
+                .parser()
+                .verifyWith(getSigningKey())
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
+    }
+}
