+++ src/main/java/gift/user/dto/request/UserModifyRequest.java
@@ -1,0 +1,15 @@
+package gift.user.dto.request;
+
+import gift.user.entity.User;
+
+public record UserModifyRequest(
+        String email,
+        String password
+) {
+    public User toEntity() {
+        return new User(
+            this.email(),
+            this.password()
+        );
+    }
+}
